refactor(cell): split space_manager/tests.rs by topic (#244)
space_manager/tests.rs was 767 lines (over the 700-line cap). Split
by SpaceManager topic into:

  tests/mod.rs              - shared TEST_SPACES_XML fixtures + make_manager helper
  tests/spaces.rs           - XML parse + space-id allocation + lookup (6 tests)
  tests/entity_lifecycle.rs - create/destroy/connect/update entity + grid sync (6 tests)
  tests/aoi.rs              - AoI enter/move/leave/re-enter + disconnect broadcast (6 tests)
  tests/witnesses.rs        - get_witnesses_of scoping (2 tests)
  tests/npc_spawn.rs        - spawn_npc + spawn_npc_from_record (5 tests)
  tests/instances.rs        - instance lifecycle + cross-instance scoping (5 tests)

`disconnect_emits_left_aoi_to_observers_then_destroys_entity` lives
under aoi.rs because the load-bearing assertion is the LeftAoI
broadcast; `find_by_tag_does_not_cross_instance_boundaries` lives
under instances.rs because the load-bearing assertion is the
instance-scoping invariant.

All 31 tests pass; no behavior change.
