+++ crates/services/src/cell/space_manager/tests/mod.rs
@@ -1,0 +1,33 @@
+//! Tests for `SpaceManager`, grouped by topic: XML/space loading, basic
+//! entity lifecycle, AoI / witness diff, witness queries, NPC spawn, and
+//! instanced-space lifecycle / scoping.
+
+use super::*;
+
+mod aoi;
+mod entity_lifecycle;
+mod instances;
+mod npc_spawn;
+mod spaces;
+mod witnesses;
+
+const TEST_SPACES_XML: &str = r#"<?xml version="1.0" encoding="UTF-8"?>
+<Spaces>
+    <Space WorldName="Agnos" Instanced="false" MinX="-2400" MaxX="2200" MinY="-3200" MaxY="2800" />
+    <Space WorldName="Castle" Instanced="false" MinX="0" MaxX="2400" MinY="0" MaxY="2400" />
+    <Space WorldName="Castle_CellBlock" Instanced="true" MinX="-800" MaxX="800" MinY="-800" MaxY="800" />
+    <Space WorldName="SGC_W1" Instanced="true" MinX="-400" MaxX="400" MinY="-400" MaxY="800" />
+</Spaces>"#;
+
+const TEST_CELL_SPACES_XML: &str = r#"<?xml version="1.0" encoding="UTF-8"?>
+<Spaces>
+    <Space WorldName="Agnos" />
+    <Space WorldName="Castle" />
+</Spaces>"#;
+
+fn make_manager() -> SpaceManager {
+    let mut mgr = SpaceManager::new(1);
+    mgr.parse_spaces_xml(TEST_SPACES_XML).unwrap();
+    mgr.create_startup_spaces(TEST_CELL_SPACES_XML).unwrap();
+    mgr
+}
